requirements.txt
asgiref==3.5.2
astroid==2.12.12
autopep8==2.0.0
certifi==2022.9.24
charset-normalizer==2.1.1
dill==0.3.6
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
gunicorn==20.1.0
idna==3.4
isort==5.10.1
lazy-object-proxy==1.8.0
mccabe==0.7.0
mysqlclient==2.1.1
numpy==1.23.5
openai==0.25.0
opencv-python==4.6.0.66
openpyxl==3.0.10
pandas==1.5.2
pandas-stubs==1.5.2.221124
Pillow==9.3.0
platformdirs==2.5.4
pycodestyle==2.9.1
pylint==2.15.5
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.64.1
types-pytz==2022.6.0.1
typing_extensions==4.4.0
urllib3==1.26.13
whitenoise==6.2.0
wrapt==1.14.1
